Fix `getPeriodForTime` util
An issue, https://github.com/canalplus/rx-player/issues/1736, noticed
that we have an issue with the `getPeriodForTime` util which links time
and Period.

The way the util was written means it had to be looped over from the
last to the first, but I don't know why I decided to replace it with a
fancier `for...of` (which goes from first to last) in a totally
unrelated refacto (https://github.com/canalplus/rx-player/pull/1507).

This broke the logic and lead in conditions that are hard to predict
infinite rebuffering issues.

I also made some other code more resilient here, because I thought that
I saw another issue first. I don't really know if that other case is
possible but I thought that it didn't cost anything to be resilient
there.