Don't destroy PeriodStream in some ambiguous "exact end time" cases
An issue, https://github.com/canalplus/rx-player/issues/1736, noticed
that we may have an infinite rebuffering issue in the following conditions:

1. The position is exactly at a given Period's end

2. The next Period is not directly consecutive to that Period: it starts
   later.

In that situation, to which Period is linked the current position:
  - the former? (as it just ends at that exact point)
  - the latter? (as it should be the Period on which we'll focus from
    now)
  - none? (technically we may argue that no Period is linked to the
    position from now on)

Different RxPlayer modules had different opinions and we encountered an
issue at least in the `StreamOrchestrator`:

  - The `Manifest` told us that this position belonged to the former
    Period (only in that specific case).

  - However the `StreamOrchestrator` thought that the position now went
    over the former Period and was now under the responsibility of the
    latter one.

We ended up creating and destroying modules linked to the former Period.

To fix this, I propose to just play nicer in the `StreamOrchestrator`
by considering that in that special scenario, we can still consider
the position as part of the former Period (in the "bound checking" code
that previously was creating the issue).