enableRepresentationAvoidance is not experimental anymore
We brought in `v4.3.0` the `enableRepresentationAvoidance` option as an
opt-in under the `experimentalOptions.enableRepresentationAvoidance`
`loadVideo` option.

I now propose to make it non-experimental anymore - and to just enable
it by default (making it opt-out), so most applications can profit from
it.

Canal+ France has been relying on it for many months now (since the
beginning of the year), the only seen effects is that the problematic
devices can now play the corresponding contents after reloading and
"avoiding" the concerned Representation.

We've not seen any false positive yet, it would be very rare - and even
if there's one it would not be dramatic to reload avoiding a media
quality for that playing session if it provoked an infinite FREEZING
when switching to it.