hotfix: do not include already-sent observation in ContentTimeBoundariesObserver
We have an issue currently (and always had yet were saved by a minor
bug fixed in our futur version) where the `ContentTimeBoundariesObserver`
would react to the initial media observation, see that the wanted
position is `0` (the initial seek just has been decided in
`initial_seek_and_play.ts` - but was done synchronously on the exact
same observation so that observation does not reflect it yet), and thus
send a `MEDIA_TIME_BEFORE_MANIFEST` warning initially.

The problem is not easy to fix:
  - should the `ContentTimeBoundariesObserver` always expect that the
    position given by the PlaybackObserver is the one we want to play?
  - should the PlaybackObserver include the concept of the initial seek?

So for now, I've done a quick fix, that still make sense: the
`ContentTimeBoundariesObserver` does not rely on previous media
observation anymore to send `MEDIA_TIME_BEFORE_MANIFEST` /
`MEDIA_TIME_AFTER_MANIFEST` as it could be stale. It only reacts to
known fresh observations.

This is not very solid though, so a better solution should be found in
the future.