rely on "provenance statements" when publishing our builds
In the light of recent [npm security news on very popular packages that
have been compromised on the npm registry](https://www.aikido.dev/blog/npm-debug-and-chalk-packages-compromised)
(we're not affected it seems), I thought about how we can improve our
publish pipeline so applications can trust it.

_(This issue began as a fishing email from an npm impersonator,
the change wasn't made on the base repo, it was only published to
the `npm` registry directly from the attacker. What it did was to
redirect cryptocurrencies transactions when in a browser)._

How it works for now
--------------------

When we publish a package on the npm registry, we do so by running
scripts (in the `scripts/` directory) locally, and then call `npm
publish` locally (while being authentified on npm with the right
token).

So applications have to trust that:
  1. what we published was indeed linked to the code present in the
     git repository at that point.
  2. that we ran the same pushed build scripts to publish it,
     without modifying files first (which could be done only locally).

How can we improve on it
------------------------

It seems that npm provides a mechanim called ["provenance
statements"](https://docs.npmjs.com/generating-provenance-statements).

The idea is to rely on a CI as a third party (they support only GitHub
and Gitlab for now) which will do both the build process and the
publishing in its entirety server-side, on a CI script that would be
easy to audit (and the auditable source files that have been pushed) by
another application wanting to rely on us.

Also, the repo linked to that CI has to be the repository listed in the
project's `package.json` (it is for us!).

What I did
----------

I updated our publish scripts so they don't do the build nor publishing
anymore.

Now they only check that conditions are right locally (e.g. the release
has not already been published etc.) and then create a version tag (both
for official and dev releases) and push to the current branch.

On those tags, CI scripts are now scheduled:

- for official releases, it builds the rx-player then publish the result
  on npm

- for dev releases, it does both the `dev` and `canal` releases. It also
  on its side that the "canal patch" is applied. That apply is checked
  multiple times (both locally and by the CI) to be sure it can be
  applied before doing any publishing.

Our build script is still complex (to be easily auditable by an
application), but I think it's a step in the right direction.

There's though still a small detail: the secret token hasn't been added
to GitHub yet (this is needed for the npm token), I'll have to contact
the right people for that and it might take some time!
I also haven't tested this, so no idea if I forgot some key step for
now.

But anyway I still commit this to be able to receive opinions on this.

Also, we did not create tags for dev releases until now, but this means
that we'll do it. I have no problem with it though, I even think that's
a good idea to be able to check more quickly what a particular dev
release contained (we would just checkout the tag).