MULTI_THREAD/DEBUG_MODE: Fix display of logs' payload
I noticed an issue on our not-yet-released future RxPlayer version where
it would not show payload of logs from our worker, when the global
`__RX_PLAYER_DEBUG_MODE__` was set to true (this unlocks our "debug mode",
it is mainly used by [RxPaired](https://github.com/canalplus/RxPaired)).

In that mode only, all worker logs are sent back to the main thread (to
group all RxPlayer logs in a common JS realm -> much easier to debug).

Since \#1717, the format of those logs changed but were not properly
handled by the main thread.

I fixed that here. There was a remaining issue for `Error` instances,
the only "complex" type that may be set as a log payload, because we
have to do serializing and deserializing steps after detecting that
a payload seems to be a serialized error (this is done with duck typing
for now). Most of the code here is linked to this situation.