CI's Pull Requests integration tests rely on the same base than our releases
To improve the readability of our release publishing script, we've put
recently a "base" file for our integration tests in a separate yml
(.github/workflows/integration_tests_base.yml) that can be called and
parametarized (for setting the right OS and browser).

We only used it for our releases workflow, not for the regular checks on
each PR yet. This PR changes that so every workflow uses it, hopefully
making those easier to maintain.

There are two important differences:

- The previous way grouped all OS under the same job. You had one for
  linux and one for windows (github-actions on macOS had too poor of a
  performance last time we checked, sadly), each browser being tested in
  each.

  Now you have 3 (browsers) * 2 (OS) different jobs instead. This should
  also make it easier to parallelize.

- We had previously a complex retrying logic after a backoff because some
of our tests are time-sensitive (rightfully, e.g. to check that playback
rate updates are not broken on an actual playback by waiting some time
and checking the position) and thus can end up with many false negatives
when the CI is not performant enough.

  I chose to remove that retrying logic for now. In any way, it should
  be doable to bring it back if it proves to be annoying.