attempt with declaration files + esbuild + swc
The RxPlayer produces several builds for each release:

1. Our es2017 build (automatically pulled by e.g. bundlers when they
   see `import` statements for the `rx-player` node module).

   It is generated as several files (one TypeScript file == one JS file,
   one [TypeScript declaration
   file](https://www.typescriptlang.org/docs/handbook/declaration-files/introduction.html)
   and one [TypeScript declaration map
   file](https://www.typescriptlang.org/tsconfig/#declarationMap) t
   have e.g. nice [LSP goto
   definition](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_definition) mapping)
   in `./dist/es2017` and is compatible to... ES2017 JavaScript

   I guess most applications rely on that one.

2. Our commonJS build (automatically pulled by e.g. bundlers on projects
   relying on `require` statements I guess).

   It is here also generated as several files (one TypeScript file = 3
   generated files, like _1_) but as
   [ES5](https://ecma-international.org/wp-content/uploads/ECMA-262_5th_edition_december_2009.pdf),
   the spec that brought `use strict`, `JSON.stringify`, `map` /
   `filter` yet before, `let` / `const` / `class` etc.

3. Our bundles, provided alongsided release notes which expose the
   RxPlayer in `window` and is ES2017.

   This one can be used by apps without complex build systems: they can
   just drop the script in a `<script>` tag and then use the global
   `window.RxPlayer`.

So the build _2_ relies on an older JS version (for legacy reasons, we
did not want to break apps to much when designing the v4 and the v3 kept
compatibility with e.g. IE11 which was not compatible to further ES
versions).

But the rest of the JS/TS ecosystem is not on the same page. TypeScript
begins to deprecate ES5 support as their output "target" and it
has been announced that [it will disappear in their future "native"
(provided as a binary - written in Go) version, their
v7](https://devblogs.microsoft.com/typescript/progress-on-typescript-7-december-2025/)

Yet for the RxPlayer, we consider that the minimum JS version supported
by our builds are part of our API and don't want to introduce any
breaking change right now, so the most straightforward step may be to
rely on another tool to keep our "commonJS build" es5-compatible.

---

There are two known tools to do just that
[`babeljs`](https://babeljs.io/) and [`swc`](https://swc.rs/). We in the
team had previous bad experiences some years ago with both (poor
documentation, poor API for a minimal Node.js build script like ours)
but as I just re-tried `swc` (I began with that one as I heard some good
about it and it advertises much faster transpilation speed than babel), I
found out that now it seemed to work with minimal work, even if I relied
on their CLI (instead of their Node.js API, that we would have
preferred but which was more awkward to use in my simple case).

So the idea is that, for the common JS build only, it now goes:
 1. `tsc` builds that in CommonJS ES2017 (to keep the same config
    between all TypeScript builds)
 2. `swc` reads each JS generated in `dist/commonjs` and re-write each
    as ES5-compatible scripts.