fix: don't replicate updates on main/ when monothreaded
This fixes the issue first seen in #1773

Until very recently, loading a content in a multitheaded or
monothreaded mode went through very different logic (declared in
different files, with distinct concepts).

To simplify the project maintainance, we decided to merge those two
into one common path.
This gave a unique logic roughly of the same complexity than the more
complex of the two (the multithreaded one), to which we added a sort of
adapter to also support the less complex (the monothreaded) one.

When doing that, we forgot about a key difference about the Manifest:

- Monothreaded mode just keeps the same shared `Manifest` instance
  everywhere.

- Multi-threaded mode uses a simple `IManifestMetadata` object on
  main thread and the more powerful `Manifest` instance only for the
  "core" part of the code, that is here running in a WebWorker.

Because our new logic is mostly the multithreaded one that we lightly
updated to support monothreaded cases, we still had some synchronization
work on that structure, for when e.g. the Manifest is updated, that is only
needed in the multithreaded scenario (no need to synchronize in the
monothreaded case as the `Manifest` instance is the same one everywhere
and thus already contain all updates immediately) even when running
monothreaded.

Worse, due to some differences between the two structures, it could then
lead to a runtime error (for now seen on Period updates in a
monothreaded case, this is issue #1773).

To fix this I did both:

- only "synchronize" updates on main thread if what we have is an
  `IManifestMetadata` object, as this is the only case where
  synchronization needs to be done.

- In the function doing the synchronization (`replicateUpdatesOnManifestMetadata`),
  I exit directly if what we have in argument is a `Manifest`
  instance, also logging an error, as mutations done in that function
  could risk breaking the assumptions of that type.

  This should be unnecessary due to the first fix, but this is to
  protect our future developments from doing the same mistake.